Restructure hashing to crypto module
