+++ src/app.module.ts
@@ -2,6 +2,7 @@ import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { MongooseModule } from '@nestjs/mongoose';
 import { AuthModule } from './auth/auth.module';
+import { CryptoModule } from './crypto/crypto.module';
 import { TicketingModule } from './ticketing/ticketing.module';
 
 @Module({
@@ -11,6 +12,7 @@ import { TicketingModule } from './ticketing/ticketing.module';
             envFilePath: ['.env.local', '.env'],
         }),
         MongooseModule.forRoot(process.env.MONGODB_URI),
+        CryptoModule,
         TicketingModule,
         AuthModule,
     ],
