+++ src/crypto/crypto.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { HashingService } from './services/hashing.service';
+
+@Module({
+    providers: [HashingService],
+    exports: [HashingService],
+})
+export class CryptoModule {}
