+++ src/ticketing/ticketing.module.ts
@@ -3,7 +3,7 @@ import { TicketsController } from './controller/tickets.controller';
 import { TicketsService } from './services/tickets.service';
 import { MongooseModule } from '@nestjs/mongoose';
 import { ticketSchema } from './services/tickets.schema';
-import { HashingService } from './services/hashing.service';
+import { HashingService } from '../crypto/services/hashing.service';
 
 @Module({
     imports: [MongooseModule.forFeature([{ name: 'Tickets', schema: ticketSchema }])],
